# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface ExpenseItem {
  id: string;
  category: string;
  name: string;
  amount: string;
  date: string;
  createdAt: string;
}

export interface ExpenseList {
  expenses: ExpenseItem[];
}

export interface CreateExpenseRequest {
  id: string;
  category: string;
  name: string;
  amount: number;
  date: string;
}

export interface BudgetItem {
  category: string;
  amount: string;
}

export interface BudgetList {
  budgets: BudgetItem[];
}

export interface UpsertBudgetRequest {
  category: string;
  amount: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};