# lib/api-zod/src/generated/types/createExpenseRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseRequest {
  id: string;
  category: string;
  name: string;
  amount: number;
  date: string;
}