# lib/api-zod/src/generated/types/expenseList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseItem } from './expenseItem';

export interface ExpenseList {
  expenses: ExpenseItem[];
}