# lib/api-zod/src/generated/types/upsertBudgetRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertBudgetRequest {
  category: string;
  amount: number;
}